Rename import controller file to index.ts
